Fix code review
