--- WhisperAPI/WhisperAPI/Services/SelectSuggestion/ISelectSuggestionService.cs
@@ -1,9 +1,0 @@
-﻿using WhisperAPI.Models;
-
-namespace WhisperAPI.Services.SelectSuggestion
-{
-    public interface ISelectSuggestionService
-    {
-        bool UpdateContextWithSelectedSuggestion(ConversationContext conversationContext, SearchQuery searchQuery);
-    }
-}
