+++ WhisperAPI/WhisperAPI/appsettings.json
@@ -1,7 +1,7 @@
 ﻿{
-  "ApiKey": "",
+  "ApiKey": "xx9d1051e9-4b41-4425-b4fe-72b698a73324",
 
-  "NlpApiBaseAddress": "http://nlpapi:5000/",
+  "NlpApiBaseAddress": "http://localhost:5000/",
 
   "MlApiBaseAddress": "http://mlapi:5000/",
